Smarter parameter extraction
Instead of numerically solving the relation between p and beta (fixing p and looking for corresponding beta), it is much simpler to decide beforehand the interesting values of beta, regularly spaced, and then claculate the corresponding values of p for which the solution is immediately given. Then use the functional definition of the posterior for p (using the updated shape parameters) to extract the precise values of the posterior we want.